make the collisions work with float variables

+++ srcs/player/collisions.c
@@ -6,7 +6,7 @@
 /*   By: wimam <walidimam69gmail.com>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/12 08:40:04 by wimam             #+#    #+#             */
-/*   Updated: 2025/09/12 08:49:54 by wimam            ###   ########.fr       */
+/*   Updated: 2025/09/12 11:34:19 by wimam            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 bool	collisions(t_cub *cub)
 {
-	int	new_x;
-	int	new_y;
+	float	new_x;
+	float	new_y;
 
 	new_x = cub->player.xp;
 	new_y = cub->player.yp;
@@ -27,7 +27,7 @@ bool	collisions(t_cub *cub)
 		new_x -= PLAYER_SPEED;
 	if (cub->key.d == true)
 		new_x += PLAYER_SPEED;
-	if (cub->parse.map[new_y][new_x] == '1')
+	if (cub->parse.map[(int) new_y][(int) new_x] == '1')
 		return (true);
 	return (false);
 }
